+++ CHANGELOG.md
@@ -157,3 +157,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - update static function to support param program id
+
+## commons [0.1.0] - PR #80
+
+### Added
+
+- Swap exact in quote